move dependency plugin interface code to a separate module
