+++ az-flow-trigger-dependency-plugin/src/main/java/flowtrigger/DependencyCheck.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package azkaban.flowtrigger;
+package flowtrigger;
 
 public interface DependencyCheck {
