+++ azkaban-web-server/src/main/java/azkaban/flowtrigger/DependencyInstance.java
@@ -16,6 +16,7 @@
 
 package azkaban.flowtrigger;
 
+import flowtrigger.DependencyInstanceContext;
 import java.util.Date;
 
 public class DependencyInstance {
