+++ azkaban-web-server/src/main/java/azkaban/flowtrigger/DependencyInstanceCallbackImpl.java
@@ -17,6 +17,8 @@
 package azkaban.flowtrigger;
 
 import com.google.common.base.Preconditions;
+import flowtrigger.DependencyInstanceCallback;
+import flowtrigger.DependencyInstanceContext;
 
 public class DependencyInstanceCallbackImpl implements DependencyInstanceCallback {
