+++ azkaban-web-server/src/main/java/azkaban/flowtrigger/DependencyPluginConfigImpl.java
@@ -17,6 +17,7 @@
 package azkaban.flowtrigger;
 
 import com.google.common.collect.ImmutableMap;
+import flowtrigger.DependencyPluginConfig;
 import java.util.Map;
 
 public class DependencyPluginConfigImpl implements DependencyPluginConfig {
