+++ azkaban-web-server/src/main/java/azkaban/flowtrigger/FlowTriggerService.java
@@ -24,6 +24,10 @@ import azkaban.project.FlowTriggerDependency;
 import azkaban.project.Project;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
+import flowtrigger.DependencyCheck;
+import flowtrigger.DependencyInstanceCallback;
+import flowtrigger.DependencyInstanceContext;
+import flowtrigger.DependencyInstanceRuntimeProps;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Collection;
