+++ azkaban-web-server/src/test/java/azkaban/flowtrigger/testplugin/TestDependencyInstanceContext.java
@@ -16,10 +16,10 @@
 
 package azkaban.flowtrigger.testplugin;
 
-import azkaban.flowtrigger.DependencyInstanceCallback;
-import azkaban.flowtrigger.DependencyInstanceConfig;
 import azkaban.flowtrigger.DependencyInstanceContext;
-import azkaban.flowtrigger.DependencyInstanceRuntimeProps;
+import flowtrigger.DependencyInstanceCallback;
+import flowtrigger.DependencyInstanceConfig;
+import flowtrigger.DependencyInstanceRuntimeProps;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
