+++ settings.gradle
@@ -25,4 +25,5 @@ include 'azkaban-exec-server'
 include 'azkaban-hadoop-security-plugin'
 include 'azkaban-solo-server'
 include 'azkaban-web-server'
+include 'az-flow-trigger-dependency-plugin'
 include 'test'
